[runtime] cleanup
